--- app/Http/Controllers/Api/CategoryController.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\Api;
-
-use App\Category;
-use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
-
-class CategoryController extends Controller
-{
-    public function index()
-    {
-        $categories = Category::get(['id', 'name']);
-        return response()->json($categories);
-    }
-}
