+++ public/js/app.js
@@ -61795,7 +61795,7 @@ var __awaiter = (undefined && undefined.__awaiter) || function (thisArg, _argume
 
 function getCategoryList() {
     return __awaiter(this, void 0, void 0, function* () {
-        const url = '/api/category/categoryList';
+        const url = '/api/top/category/categoryList';
         const result = yield axios__WEBPACK_IMPORTED_MODULE_0___default.a
             .get(url)
             .then((response) => {
