+++ public/mix-manifest.json
@@ -1,4 +1,4 @@
 {
-    "/js/app.js": "/js/app.js?id=be52f5609ad4c4ba3c0f",
+    "/js/app.js": "/js/app.js?id=99ce1a37f415c3981a26",
     "/css/app.css": "/css/app.css?id=6332612009a93cd6a053"
 }
