+++ routes/api.php
@@ -15,16 +15,15 @@ use Illuminate\Support\Facades\Route;
 */
 
 Route::middleware('auth:api')->get('/user', function (Request $request) {
-    return $request->user();
+  return $request->user();
 });
 
 Route::group(['middleware' => ['api']], function() {
-    Route::resource('category/categoryList', 'Api\CategoryController', ['except' => ['create', 'edit']]);
+  Route::prefix('top')->group(function() {
+    Route::get('/latestList', 'Api\TopController@latest');
+    Route::get('/category/categoryList', 'Api\TopController@category');
+  });
 
-    Route::prefix('top')->group(function() {
-        Route::get('/latestList', 'Api\TopController@latest');
-    });
-
-    Route::post('ronbun/{id}', 'Api\TopController@show');
-    Route::post('/category/{category_id}/ronbunList', 'Api\TopController@category');
+  Route::post('/ronbun/{id}', 'Api\RonbunController@index');
+  Route::post('/category/{category_id}/ronbunList', 'Api\RonbunListController@category');
 });
