+++ tests/Http/Controllers/Api/TopControllerTest.php
@@ -84,56 +84,6 @@ class TopControllerTest extends TestCase
     $this->assertEquals('テストカテゴリー', $ronbuns->first()->category_name);
   }
 
-  /** @test */
-  public function 指定したIDの論文情報が取得できる()
-  {
-    $category = $this->createCateogry();
-    $ronbun1 = $this->createRonbun(
-      $title = 'テストタイトル1',
-      $author = 'テスト著者1',
-      $abstract = 'テストテキスト1テストテキスト1テストテキスト1テストテキスト1',
-      $category_id = $category->id,
-    );
-    $ronbun2 = $this->createRonbun(
-      $title = 'テストタイトル2',
-      $author = 'テスト著者2',
-      $abstract = 'テストテキスト2テストテキスト2テストテキスト2テストテキスト2',
-      $category_id = $category->id,
-    );
-
-    $ronbun = (new TopController)->getRonbun($ronbun1->id);
-
-    $this->assertEquals('テストタイトル1', $ronbun->title);
-  }
-
-  /** @test */
-  public function 指定したカテゴリーIDに紐づく論文一覧が取得できる()
-  {
-    $category = $this->createCateogry();
-    $this->createRonbun(
-      $title = 'テストタイトル1',
-      $author = 'テスト著者1',
-      $abstract = 'テストテキスト1テストテキスト1テストテキスト1テストテキスト1',
-      $category_id = $category->id,
-    );
-    $this->createRonbun(
-      $title = 'テストタイトル2',
-      $author = 'テスト著者2',
-      $abstract = 'テストテキスト2テストテキスト2テストテキスト2テストテキスト2',
-      $category_id = $category->id,
-    );
-    $this->createRonbun(
-      $title = 'テストタイトル3',
-      $author = 'テスト著者3',
-      $abstract = 'テストテキスト3テストテキスト3テストテキスト3テストテキスト3',
-      $category_id = 1,
-    );
-
-    $ronbunList = (new TopController)->getCategoryRonbunList($category->id);
-
-    $this->assertEquals(2, $ronbunList->count());
-  }
-
   private function createRonbun($title, $author, $abstract, $category_id = 1)
   {
     return Ronbun::create([
